+++ lib/pod/command/plugins.rb
@@ -20,6 +20,7 @@ module Pod
         Shows the available CocoaPods plugins and if you have them installed or not.
       DESC
 
+      # Force-download the JSON
       def download_json
         response = REST.get(PLUGINS_URL)
         if response.ok?
@@ -27,6 +28,7 @@ module Pod
         end
       end
 
+      # Tells if a gem is installed
       def installed?(gemname)
         if Gem::Specification.methods.include?(:find_all_by_name)
           Gem::Specification.find_all_by_name(gemname).any?
@@ -36,7 +38,8 @@ module Pod
         end
       end
 
-      def run
+      # Download the JSON if not already in @json, then execute the given block
+      def fetch_plugins_list
         UI.puts 'Downloading Plugins list...'
         begin
           download_json unless json
@@ -45,28 +48,88 @@ module Pod
         end
 
         if json
-          print_plugins
+          yield @json['plugins'] if block_given?
         else
           UI.puts 'Could not download plugins list from cocoapods.org'
         end
       end
 
-      def print_plugins
-        UI.puts "Available CocoaPods Plugins\n\n"
+      def run
+        fetch_plugins_list do |plugins|
+          UI.puts "\nAvailable CocoaPods Plugins\n"
 
-        @json['plugins'].each do |plugin|
-
-          plugin_name = "-> #{plugin['name']}"
-          plugin_colored_name = installed?(plugin['gem']) ? plugin_name.green : plugin_name.yellow
-
-          UI.title(plugin_colored_name, '', 1) do
-            UI.puts_indented plugin['description']
-            UI.labeled('Gem', plugin['gem'])
-            UI.labeled('URL',   plugin['url'])
-            UI.labeled('Author', plugin['author']) if self.verbose?
+          plugins.each do |plugin|
+            print_plugin plugin
           end
-
         end
+      end
+
+      # Display information about a plugin given its hash
+      def print_plugin(plugin)
+        plugin_name = "-> #{plugin['name']}"
+        plugin_colored_name = installed?(plugin['gem']) ? plugin_name.green : plugin_name.yellow
+
+        UI.title(plugin_colored_name, '', 1) do
+          UI.puts_indented plugin['description']
+          UI.labeled('Gem', plugin['gem'])
+          UI.labeled('URL',   plugin['url'])
+          UI.labeled('Author', plugin['author']) if self.verbose?
+        end
+      end
+
+      #-----------------------------------------------------------------------#
+
+      # The search subcommand. Used to search a plugin in the list of known plugins,
+      # searching into the name and description fields
+      #
+      class Search < Plugins
+        self.summary = 'Search for known plugins'
+
+        self.description = <<-DESC
+          Search plugins whose name contains the given text (ignoring case).
+          With --full, search by name but also author and description
+        DESC
+
+        self.arguments = 'QUERY'
+
+        def self.options
+          [
+            ["--full",  "Search by name, author, and description"],
+          ].concat(super.reject { |option, _| option == '--silent' })
+        end
+
+        def initialize(argv)
+          @full_text_search = argv.flag?('full')
+          @query = argv.shift_argument unless argv.arguments.empty?
+          super
+        end
+
+        def validate!
+          super
+          help! "A search query is required." if @query.nil? || @query.empty?
+          begin
+            /#{@query}/
+          rescue RegexpError
+            help! "A valid regular expression is required."
+          end
+        end
+
+        def run
+          fetch_plugins_list do |plugins|
+            UI.puts "\nAvailable CocoaPods Plugins matching '#{@query}'\n"
+
+            query_regexp = /#{@query}/i
+            plugins.each do |plugin|
+              texts = [plugin['name']]
+              if @full_text_search
+                texts << plugin['author'] if plugin['author']
+                texts << plugin['description'] if plugin['description']
+              end
+              print_plugin plugin unless texts.grep(query_regexp).empty?
+            end
+          end
+        end
+
       end
 
       #-----------------------------------------------------------------------#